Migrate test-unit-transform.py to pytest
